feat(cost): wire CostGuard + RateLimiter, money as integer micro-USD
All cost / pricing / budget values now flow through a `MicroUsd(i64)`
newtype in `aura-model`; SQL columns store INTEGER, the analytics wire
format uses `cost_micro_usd: int64`, and TOML still accepts USD floats
(`user_daily_usd = 5.0`) via a serde-with bridge — so quota arithmetic
stays exact while operator-facing surfaces stay readable. Runtime now
constructs `SpendingLimits` + `CostGuard` from the config block and
attaches both that and the configured rate limiter to the `Router`, so
caps in `cost.spending_limits` actually gate requests.

+++ crates/storage/src/libsql/mod.rs
@@ -162,7 +162,10 @@ impl LibsqlPool {
                     output_tokens                   INTEGER NOT NULL,
                     cached_input_tokens             INTEGER NOT NULL DEFAULT 0,
                     cache_creation_input_tokens     INTEGER NOT NULL DEFAULT 0,
-                    cost_usd                        REAL    NOT NULL,
+                    -- Spend in micro-USD (USD × 10^6). INTEGER, never REAL —
+                    -- floats accumulate rounding error across SUM() and
+                    -- quota comparisons.
+                    cost_usd                        INTEGER NOT NULL,
                     timestamp                       INTEGER NOT NULL
                 );
                 CREATE INDEX IF NOT EXISTS idx_cost_user_id ON cost_records(user_id);
@@ -173,7 +176,8 @@ impl LibsqlPool {
                 CREATE TABLE IF NOT EXISTS user_monthly_cost (
                     user_id     TEXT    NOT NULL,
                     month       TEXT    NOT NULL,
-                    cost_usd    REAL    NOT NULL,
+                    -- Cumulative spend in micro-USD; see cost_records above.
+                    cost_usd    INTEGER NOT NULL,
                     updated_at  INTEGER NOT NULL,
                     PRIMARY KEY (user_id, month)
                 );
